+++ trunk/src/dll/udefrag/map.c
@@ -246,7 +246,7 @@ int is_mft(winx_file_info *f,udefrag_job_parameters *jp)
     if(is_not_mft_file(f)) return 0;
     if(is_mft_file(f)) return 1;
     
-    length = wcslen(f->path);
+    length = (int)wcslen(f->path);
     if(length == 11){
         if(winx_wcsistr(f->name,mft_name)){
             f->user_defined_flags |= UD_FILE_MFT_FILE;
